Add io_uring TLS transfer for Linux and update docs
Introduces io_uring-based TLS bidirectional transfer for Linux 5.1+ in src/io/tls_transfer/transfer.zig and src/io/tls_transfer/uring.zig, providing faster event notification for TLS connections. Updates USAGE.md with practical SSL certificate generation, file transfer, and remote shell examples. Adds BUILD.md with detailed build and packaging instructions. Improves README.md for clarity and brevity. Minor logging API changes and bugfixes in src/client.zig. Other files updated for io_uring support and related refactoring.

+++ src/server/broker/protocols.zig
@@ -2,7 +2,9 @@ const std = @import("std");
 const broker = @import("../broker.zig");
 const logging = @import("../../util/logging.zig");
 
-pub fn initializeChatClient(self: *broker.BrokerServer, client_id: u32) !void {
+/// Initialize a new chat client with welcome message and nickname prompt
+/// SECURITY FIX (2025-10-10): Changed client_id from u32 to u64 to match ClientPool
+pub fn initializeChatClient(self: *broker.BrokerServer, client_id: u64) !void {
     const client = self.clients.getClient(client_id) orelse return broker.BrokerError.ClientNotFound;
 
     const welcome_msg = "Welcome to ZigCat chat! Please enter your nickname: ";
